Music Connect

+++ sb1024_test/src/main/java/edu/du/sb1024/controller/AdminController.java
@@ -4,13 +4,11 @@ import edu.du.sb1024.entity.Member;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.ModelAttribute;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.*;
 
 import javax.persistence.EntityManager;
 import javax.persistence.EntityManagerFactory;
+import java.util.ArrayList;
 import java.util.List;
 
 @Controller
@@ -18,27 +16,47 @@ import java.util.List;
 public class AdminController {
     final EntityManagerFactory emf;
 
-    @GetMapping("/admin")
+    @GetMapping("/admin")//권한이 유저인(실제 사용자)의 정보를 모두 확인할 수 있음. 아마 생성 순서(id 순서)에 따라 호출
     public String admin(Model model) {
         EntityManager em = emf.createEntityManager();
-        List<Member> mems = em.createQuery("select m from Member m where m.role = :role order by m.id desc", Member.class).setParameter("role","user").getResultList();
-        model.addAttribute("members", mems);
-        return "/유저관리 페이지 ";
+        List<Member> members = em.createQuery("select m from Member m where m.role = :role order by m.id desc", Member.class).setParameter("role", "user").getResultList();
+        model.addAttribute("members", members);
+        em.close();
+        return "/admin/management";
     }
-//    @PostMapping("/admin")
+
+    //    @PostMapping("/admin")
 //    public String show(Model model) {
 //        EntityManager em = emf.createEntityManager();
 //        List<Member> mems = em.createQuery("select m from Member m where m.role = :role order by m.id desc", Member.class).setParameter("role","user").getResultList();
 //        model.addAttribute("members", mems);
 //        return "/유저관리 페이지";
 //    }
-    @PostMapping("/admin/delete/{id}")
-    public String delete(@PathVariable Long id){
+    @PostMapping("/admin/delete/{id}")//각각의 id 값을 받아서 삭제
+    public String delete(@PathVariable Long id) {
         EntityManager em = emf.createEntityManager();
         em.getTransaction().begin();
         Member member = em.find(Member.class, id);
-        em.remove(member);
+        if (member != null) {
+            em.remove(member);
+        }
         em.getTransaction().commit();
+        em.close();
         return "redirect:/admin";
+    }
+
+    @PostMapping("/search/{keyword}")//검색창을 만들어 이름, 별명, 이메일등을 카테고리로 하여 검색함.
+    public String search(@PathVariable String keyword, @RequestParam("search") String search, @ModelAttribute("members") Member member, Model model) {
+        EntityManager em = emf.createEntityManager();
+        List<Member> members = new ArrayList<>();
+        if (keyword.equals("nick")) {
+            members = em.createQuery("select m from Member m where m.nick like :search", Member.class).setParameter("search", search).getResultList();
+        } else if (keyword.equals("name")) {
+            members = em.createQuery("select m from Member m where m.username like :search", Member.class).setParameter("search", search).getResultList();
+        } else if (keyword.equals("email")) {
+            members = em.createQuery("select m from Member m where m.email like :search", Member.class).setParameter("search", search).getResultList();
+        }
+        model.addAttribute("members", members);
+        return "/admin/management";
     }
 }
